api/api/requirements.txt
tensorflow==2.5.0
fastapi
uvicorn
python-multipart
pillow
tensor-flow-serving-api==2.5.0
matplotlib
numpy